Unit Attack
Fix: HpBar Update

+++ Source/URR/UI/URRHpBarWidget.h
@@ -28,6 +28,6 @@ protected:
 	void UpdateHpBar();
 
 protected:
-	int CurrentHealth;
-	int CurrentMaxHealth;
+	float CurrentHealth;
+	float CurrentMaxHealth;
 };
